+++ tests/Datasets/UnlabeledTest.php
@@ -207,22 +207,6 @@ class UnlabeledTest extends TestCase
         $this->assertCount(3, $subset);
     }
 
-    public function test_save_and_load()
-    {
-        $this->assertFalse(file_exists(__DIR__ . '/test.dataset'));
-
-        $this->dataset->save(__DIR__ . '/test.dataset');
-
-        $this->assertFileExists(__DIR__ . '/test.dataset');
-
-        $dataset = Unlabeled::load(__DIR__ . '/test.dataset');
-
-        $this->assertInstanceOf(Unlabeled::class, $dataset);
-        $this->assertInstanceOf(Dataset::class, $dataset);
-
-        $this->assertTrue(unlink(__DIR__ . '/test.dataset'));
-    }
-
     public function test_merge_dataset()
     {
         $this->assertCount(count($this->samples), $this->dataset);