+++ tests/Regressors/AdalineTest.php
@@ -29,7 +29,7 @@ class AdalineTest extends TestCase
 
     public function setUp()
     {
-        $this->training = Labeled::load(dirname(__DIR__) . '/mpg.dataset');
+        $this->training = unserialize(file_get_contents(dirname(__DIR__) . '/mpg.dataset') ?: '');
 
         $this->testing = $this->training->randomize()->head(3);